Add job that fetches electricity prices
