+++ src/event_loop/optimization_job.rs
@@ -1,4 +1,5 @@
 use super::arrow_input;
+use super::electricity_price_job;
 use super::job_store::JobStore;
 use super::jobs::{JobOutcome, JobStatus, OptimizationOutcome};
 use super::time_series;
@@ -19,7 +20,6 @@ use arrow::record_batch::RecordBatch;
 use arrow_ipc::reader::StreamReader;
 use chrono::{DateTime, DurationRound, TimeDelta, Utc};
 use juniper::GraphQLObject;
-use serde_json::Value;
 use std::collections::BTreeMap;
 use std::error::Error;
 use std::process::{Command, ExitStatus};
@@ -770,117 +770,6 @@ pub async fn fetch_weather_data_task(
     }
 }
 
-async fn fetch_electricity_prices(
-    start_time: String,
-    end_time: String,
-    country: &String,
-) -> Result<Vec<(TimeStamp, f64)>, Box<dyn std::error::Error + Send + Sync>> {
-    let client = reqwest::Client::new();
-    let url = format!(
-        "https://dashboard.elering.ee/api/nps/price?start={}&end={}",
-        start_time, end_time
-    );
-    let response = client
-        .get(&url)
-        .header("accept", "application/json")
-        .send()
-        .await
-        .map_err(|e| Box::new(e) as Box<dyn std::error::Error + Send + Sync>)?;
-    let response_text = response
-        .text()
-        .await
-        .map_err(|e| Box::new(e) as Box<dyn std::error::Error + Send + Sync>)?;
-    Ok(parse_elering_response(
-        &response_text,
-        &as_elering_country(country)?,
-    )?)
-}
-
-fn parse_elering_response(
-    response_text: &str,
-    elering_country: &str,
-) -> Result<Vec<(TimeStamp, f64)>, String> {
-    let response = serde_json::from_str(response_text)
-        .or_else(|error| Err(format!("failed to parse response: {}", error)))?;
-    let response_object = match response {
-        Value::Object(object) => object,
-        _ => return Err("unexpected response content".to_string()),
-    };
-    let success = match response_object
-        .get("success")
-        .ok_or("'success' field missing in response".to_string())?
-    {
-        Value::Bool(success) => success,
-        _ => return Err("unexpected type for 'success' in response".to_string()),
-    };
-    if !success {
-        return Err("electricity price query unsuccessful".to_string());
-    }
-    let data = match response_object
-        .get("data")
-        .ok_or("'data' field missing in response".to_string())?
-    {
-        Value::Object(data) => data,
-        _ => return Err("unexpected type for 'data' in response".to_string()),
-    };
-    let price_data = match data.get(elering_country).ok_or(format!(
-        "requested country '{}' not found in response",
-        elering_country
-    ))? {
-        Value::Array(price_data) => price_data,
-        _ => return Err("unexpected type for country data in response".to_string()),
-    };
-    let mut pairs = Vec::with_capacity(price_data.len());
-    for element in price_data {
-        pairs.push(elering_time_stamp_pair(element)?);
-    }
-    Ok(pairs)
-}
-
-fn as_elering_country(country: &str) -> Result<&str, String> {
-    if country == "Estonia" {
-        return Ok("ee");
-    }
-    if country == "Finland" {
-        return Ok("fi");
-    }
-    if country == "Lithuania" {
-        return Ok("lt");
-    }
-    if country == "Latvia" {
-        return Ok("lv");
-    }
-    Err("unknown or unsupported country".to_string())
-}
-
-fn elering_time_stamp_pair(time_stamp_entry: &Value) -> Result<(TimeStamp, f64), String> {
-    let stamp_object = match time_stamp_entry {
-        Value::Object(stamp_object) => stamp_object,
-        _ => return Err("unexpected type for price entry in response".to_string()),
-    };
-    let time_stamp = match stamp_object
-        .get("timestamp")
-        .ok_or("'timestamp' field missing in price entry")?
-    {
-        Value::Number(t) => t
-            .as_i64()
-            .ok_or("failed to parse time stamp as integer".to_string())?,
-        _ => return Err("unexpected type for time stamp".to_string()),
-    };
-    let time_stamp = DateTime::from_timestamp(time_stamp, 0)
-        .ok_or("failed to convert time stamp to date time".to_string())?;
-    let price = match stamp_object
-        .get("price")
-        .ok_or("'price' field missing in price entry")?
-    {
-        Value::Number(x) => x
-            .as_f64()
-            .ok_or("failed to parse price as float".to_string())?,
-        _ => return Err("unexpected type for price".to_string()),
-    };
-    Ok((time_stamp, price))
-}
-
 fn create_and_update_elec_price_data(
     optimization_data: &mut OptimizationData,
     price_series: &BTreeMap<TimeStamp, f64>,
@@ -949,14 +838,17 @@ async fn fetch_electricity_price_task(
             .time_data
             .as_ref()
             .ok_or("fetch_electricity_price_task: didn't receive time data")?;
-        let format_string = "%Y-%m-%dT%H:%M:%S%.3fZ";
         let start_time = time_line.first().ok_or("time line is empty")?;
-        let start_time_str = start_time.format(&format_string).to_string();
         let end_time = (*time_line.last().ok_or("time line is empty")? - TimeDelta::hours(1))
             .duration_trunc(TimeDelta::hours(1))
             .unwrap();
-        let end_time_str = end_time.format(&format_string).to_string();
-        match fetch_electricity_prices(start_time_str, end_time_str, &location.country).await {
+        match electricity_price_job::fetch_electricity_prices(
+            &location.country,
+            start_time,
+            &end_time,
+        )
+        .await
+        {
             Ok(prices) => {
                 let prices = fit_prices_to_time_line(&prices, time_line)?;
                 if let Err(error) =
@@ -1276,55 +1168,6 @@ mod tests {
                 process_names[0].1,
                 "control_process_input_node_control_process"
             );
-        }
-    }
-    mod as_elering_country {
-        use super::*;
-        #[test]
-        fn gives_correct_country_codes() -> Result<(), String> {
-            assert_eq!(as_elering_country("Estonia")?, "ee");
-            assert_eq!(as_elering_country("Finland")?, "fi");
-            assert_eq!(as_elering_country("Lithuania")?, "lt");
-            assert_eq!(as_elering_country("Latvia")?, "lv");
-            Ok(())
-        }
-        #[test]
-        fn fails_with_unknown_country() {
-            if let Ok(..) = as_elering_country("Mordor") {
-                panic!("call should have failed");
-            }
-        }
-    }
-    mod elering_time_stamp_pair {
-        use super::*;
-        use chrono::TimeZone;
-        #[test]
-        fn parses_time_stamp_entry() {
-            let entry_json = "{\"timestamp\":1731927600,\"price\":85.0600}";
-            let entry = serde_json::from_str(entry_json).expect("entry JSON should be parseable");
-            let pair =
-                elering_time_stamp_pair(&entry).expect("constructing a pair should not fail");
-            let expected_date_time = Utc.with_ymd_and_hms(2024, 11, 18, 11, 00, 00).unwrap();
-            assert_eq!(pair, (expected_date_time, 85.06));
-        }
-    }
-    mod parse_elering_response {
-        use super::*;
-        use chrono::TimeZone;
-        #[test]
-        fn parses_response_succesfully() {
-            let response_json = r#"
-{
-    "success":true,
-    "data": {
-        "fi":[{"timestamp":1731927600,"price":70.0500}]
-    }
-}"#;
-            let time_series =
-                parse_elering_response(response_json, "fi").expect("parsing should succeed");
-            let expected_date_time = Utc.with_ymd_and_hms(2024, 11, 18, 11, 00, 00).unwrap();
-            let expected_time_series = vec![(expected_date_time, 70.05)];
-            assert_eq!(time_series, expected_time_series);
         }
     }
     mod fit_prices_to_time_line {
